Migrate integrations out of admin namespace
