+++ app/views/integrations/edit.haml
@@ -5,12 +5,12 @@
   .container-fluid
     .row-fluid
       = render 'form'
-      = link_to "Delete Integration", admin_integration_path(@integration), class: 'btn btn-danger', method: :delete, data: {confirm: "Are you sure?"}
+      = link_to "Delete Integration", integration_path(@integration), class: 'btn btn-danger', method: :delete, data: {confirm: "Are you sure?"}
     %hr
     - if @integration.failures.count > 0
       %h4
         Failures
-        = link_to "(Clear All)", clear_failures_admin_integration_path(@integration), method: :post
+        = link_to "(Clear All)", clear_failures_integration_path(@integration), method: :post
       .table-responsive
         %table.table
           %thead
