+++ app/views/integrations/new.haml
