+++ app/views/peek/views/_integration_failures.haml
@@ -1,5 +1,5 @@
 .peek-link.hidden-xs
-  = link_to admin_integrations_path do
+  = link_to integrations_path do
     %strong= view.failure_count
     %i.fa.fa-comment
     failed integrations
